Phase 3: Extract auth layer into src/auth.py (~340 lines, 10 tests)
Extracted from app.py:
- show_login_page()   — login / register / guest UI
- check_authentication() — session gate + periodic session validation
- show_announcements() / get_announcements() — used from login page and main()
- ENABLE_SESSION_MANAGER / _session_manager — session-cookie infrastructure
- on_logout() — new thin helper; replaces direct _session_manager calls in
  the logout button handler so app.py has no knowledge of the session manager

app.py changes:
- Imports init_session_manager, check_authentication, show_announcements,
  on_logout from auth
- Logout button calls on_logout() instead of touching _session_manager directly
- ~340 lines removed (2808 → 2466)

Tests: 10 new in tests/test_auth.py (import smoke + on_logout safety)
All 47 tests pass.
